ogl: move duplicate methods to panel class
